+++ server/click/db.go
@@ -1,4 +1,4 @@
-package main
+package click
 
 import "sync"
